video_to_srt/pyproject.toml
[project]
name = "video-to-srt"
version = "0.1.0"
description = "AI-powered video to SRT subtitle generation using DashScope API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dashscope>=1.13.3",
    "moviepy>=1.0.3",
    "requests>=2.31.0",
    "oss2>=2.17.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
exclude = [
    "input_videos",
    "output_srt", 
    "video_srt_env",
    "*.txt",
    "__pycache__",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"

[project.scripts]
video-to-srt = "run:main"
batch-video-to-srt = "batch_video_to_srt:main"
